chore(cyclonedx): Split functions across files
Improve overview by putting all `Bom` object related extension functions
into a separate file, and other utility functions into another file.
